Prueba

+++ api.py
@@ -1,22 +1,16 @@
 from fastapi import FastAPI, Response, status
 import pymysql.cursors
 import json
-
+import credentials
 
 app = FastAPI()
 
 @app.get("/getCharolas")
 async def root(response: Response):
     try:
-        connection = pymysql.connect(
-            host='proyectosfcqi.tij.uabc.mx',
-            port=3306,
-            user='becerra20242',
-            password='2532',
-            db='bd1becerra20242'
-        )
+        connection = credentials.get_connection()
         with connection.cursor() as cursor:
-            cursor.execute("SELECT * FROM Charola")
+            cursor.execute("SELECT * FROM charola")
             result = cursor.fetchall()
             if not result:
                 response.status_code = status.HTTP_404_NOT_FOUND
@@ -30,13 +24,8 @@ async def root(response: Response):
 @app.get("/getCharola/{index}")
 async def root(index : int, response: Response):
     try:
-        connection = pymysql.connect(
-            host='proyectosfcqi.tij.uabc.mx',
-            port=3306,
-            user='becerra20242',
-            password='2532',
-            db='bd1becerra20242'
-        )
+        connection = credentials.get_connection()
+
         with connection.cursor() as cursor:
             cursor.execute("SELECT * FROM Instrumento WHERE idCharola = %s", (index))
             result = cursor.fetchall()
@@ -52,13 +41,8 @@ async def root(index : int, response: Response):
 @app.get("/getDepartamento/{index}")
 async def root(index : int, response: Response):
     try:
-        connection = pymysql.connect(
-            host='proyectosfcqi.tij.uabc.mx',
-            port=3306,
-            user='becerra20242',
-            password='2532',
-            db='bd1becerra20242'
-        )
+        connection = credentials.get_connection()
+
         with connection.cursor() as cursor:
             cursor.execute("SELECT * FROM Departamento WHERE idDepartamento = %s", (index))
             result = cursor.fetchall()
@@ -76,13 +60,8 @@ async def root(index : int, response: Response):
 @app.put("/updateUsuario/{index}/{campo}/{nuevoValor}")
 async def root(index : int, campo : str , nuevoValor : str , response: Response):
     try:
-        connection = pymysql.connect(
-            host='proyectosfcqi.tij.uabc.mx',
-            port=3306,
-            user='becerra20242',
-            password='2532',
-            db='bd1becerra20242'
-        )
+        connection = credentials.get_connection()
+
         with connection.cursor() as cursor:
             query = f"UPDATE Persona SET {campo} = %s WHERE idPersona = %s"
             cursor.execute(query, (nuevoValor, index))
@@ -98,13 +77,8 @@ async def root(index : int, campo : str , nuevoValor : str , response: Response)
 @app.get("/getUsuarios")
 async def root(response: Response):
     try:
-        connection = pymysql.connect(
-            host='proyectosfcqi.tij.uabc.mx',
-            port=3306,
-            user='becerra20242',
-            password='2532',
-            db='bd1becerra20242'
-        )
+        connection = credentials.get_connection()
+
         with connection.cursor() as cursor:
             cursor.execute("SELECT * FROM Persona")
             result = cursor.fetchall()
@@ -122,13 +96,8 @@ async def root(response: Response):
 @app.get("/getUsuario/{index}")
 async def root(index : int, response: Response):
     try:
-        connection = pymysql.connect(
-            host='proyectosfcqi.tij.uabc.mx',
-            port=3306,
-            user='becerra20242',
-            password='2532',
-            db='bd1becerra20242'
-        )
+        connection = credentials.get_connection()
+
         with connection.cursor() as cursor:
             cursor.execute("SELECT * FROM Persona WHERE idPersona = %s", (index))
             result = cursor.fetchall()
@@ -146,13 +115,8 @@ async def root(index : int, response: Response):
 @app.post("/postUsuario/{nombre}/{Rol}/{Departamento_idDepartamento}")
 async def root(nombre : str, Rol : str, Departamento_idDepartamento : int, response: Response):
     try:
-        connection = pymysql.connect(
-            host='proyectosfcqi.tij.uabc.mx',
-            port=3306,
-            user='becerra20242',
-            password='2532',
-            db='bd1becerra20242'
-        )
+        connection = credentials.get_connection()
+
         with connection.cursor() as cursor:
             cursor.execute("INSERT INTO Persona (nombre, Rol, Departamento_idDepartamento) VALUES (%s, %s, %s)", (nombre, Rol, Departamento_idDepartamento))
             connection.commit()
@@ -167,13 +131,8 @@ async def root(nombre : str, Rol : str, Departamento_idDepartamento : int, respo
 @app.delete("/deleteUsuario/{index}")
 async def root(index : int, response: Response):
     try:
-        connection = pymysql.connect(
-            host='proyectosfcqi.tij.uabc.mx',
-            port=3306,
-            user='becerra20242',
-            password='2532',
-            db='bd1becerra20242'
-        )
+        connection = credentials.get_connection()
+
         with connection.cursor() as cursor:
             cursor.execute("DELETE FROM Persona WHERE idPersona = %s", (index))
             connection.commit()
@@ -188,13 +147,8 @@ async def root(index : int, response: Response):
 @app.get("/getEspecialidades")
 async def root(response: Response):
     try:
-        connection = pymysql.connect(
-            host='proyectosfcqi.tij.uabc.mx',
-            port=3306,
-            user='becerra20242',
-            password='2532',
-            db='bd1becerra20242'
-        )
+        connection = credentials.get_connection()
+
         with connection.cursor() as cursor:
             cursor.execute("SELECT * FROM Departamento")
             result = cursor.fetchall()
@@ -212,13 +166,8 @@ async def root(response: Response):
 @app.post("/postEspecialidad/{nombre}/{descripcion}")
 async def root(nombre : str, descripcion : str, response: Response):
     try:
-        connection = pymysql.connect(
-            host='proyectosfcqi.tij.uabc.mx',
-            port=3306,
-            user='becerra20242',
-            password='2532',
-            db='bd1becerra20242'
-        )
+        connection = credentials.get_connection()
+
         with connection.cursor() as cursor:
             cursor.execute("INSERT INTO Departamento (nombre, descripcion) VALUES (%s, %s)", (nombre, descripcion))
             connection.commit()
@@ -233,18 +182,19 @@ async def root(nombre : str, descripcion : str, response: Response):
 @app.delete("/deleteEspecialidad/{index}")
 async def root(index : int, response: Response):
     try:
-        connection = pymysql.connect(
-            host='proyectosfcqi.tij.uabc.mx',
-            port=3306,
-            user='becerra20242',
-            password='2532',
-            db='bd1becerra20242'
-        )
+        connection = credentials.get_connection()
+
         with connection.cursor() as cursor:
-            cursor.execute("DELETE FROM Departamento WHERE idDepartamento = %s", (index))
-            connection.commit()
-            response.status_code = status.HTTP_200_OK
-            return {"message": "Especialidad eliminada correctamente"}
+            cursor.execute("SELECT * FROM Departamento WHERE idDepartamento = %s", (index))
+            result = cursor.fetchall()
+            if not result:
+                response.status_code = status.HTTP_404_NOT_FOUND
+                return {"message": "No existe la especialidad"}
+            else:
+                cursor.execute("DELETE FROM Departamento WHERE idDepartamento = %s", (index))
+                connection.commit()
+                response.status_code = status.HTTP_200_OK
+                return {"message": "Especialidad eliminada correctamente"}
     except Exception as e:
         error = "Error: " + str(e)
         return error
